requirements.txt
# Deep Learning Framework
pytorchvideo>=0.1.5                    
segmentation-models-pytorch>=0.5.0                    
torch>=2.4.1                    
torchmetrics>=1.8.1                    
torchseg>=0.0.1a4                  
torchvision>=0.19.1
# Web Framework
Django>=5.2.5
python-dotenv>=1.0.0

# Video Processing
opencv-python>=4.8.0
moviepy>=1.0.3
numpy>=1.24.0
Pillow>=10.0.0

# Development Tools
jupyter>=1.0.0
notebook>=7.0.0
ipython>=8.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# File Handling
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system == "Windows"

# Testing
pytest>=7.4.0
pytest-django>=4.5.2

# Documentation
Sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Security
django-cors-headers>=4.2.0
django-cleanup>=8.0.0

# Deployment
gunicorn>=21.2.0; platform_system != "Windows"
whitenoise>=6.5.0
